pyproject.toml
[tool.poetry]
name = "anverali-test"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
